requirements.txt
pandas
torch
sentence-transformers
scikit-learn
faiss-cpu
numpy
dill